backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "duckduckgo-search>=8.0.2",
    "exa-py>=1.13.1",
    "faker>=37.3.0",
    "fastapi[standard]>=0.115.12",
    "forex-python>=1.9.2",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-storage>=3.1.0",
    "greenlet>=3.2.2",
    "langchain-chroma>=0.2.3",
    "langchain-experimental>=0.3.4",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
    "langgraph-checkpoint-postgres>=2.0.21",
    "linkup-sdk>=0.2.6",
    "markdownify>=1.1.0",
    "matplotlib>=3.10.3",
    "mftool>=2.9",
    "nselib>=1.9",
    "pandas-market-calendars>=5.1.0",
    "passlib>=1.7.4",
    "pdfminer-six>=20250327",
    "pdfplumber>=0.11.6",
    "psycopg2>=2.9.10",
    "psycopg[binary,pool]>=3.2.9",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "redis>=6.1.0",
    "selenium>=4.32.0",
    "sqlalchemy>=2.0.41",
    "tavily-python>=0.7.3",
    "yfinance>=0.2.61",
    "youtube-search>=2.1.2",
]
